+++ lib/src/enums/tournament_rank.dart
@@ -1,24 +1,38 @@
 import 'dart:math';
 
+import 'package:bridge_game/src/enums/scoring_and_tournament_type.dart';
+
 enum TournamentRank {
-  otx_4x('OTX****', 15, 300),
-  otx_3x('OTX***', 12, 200),
-  otx_2x('OTX**', 10, 100),
-  otx('OTX', 7, 50),
-  otxUnder40('OTX <40 boards', 5, 0),
-  rtx_1x('RTX*', 5, 50),
-  rtx('RTX', 5, 0),
-  rtxUnder40('RTX <40 boards', 3, 0),
-  okrTx('Okr. TX', 2, 0),
-  clubTx('Klub. TX', 1, 0);
+  otx_4x('OTX****', 25, 200, 100, 40, 300, 100),
+  otx_3x('OTX***', 15, 150, 100, 25, 200, 100),
+  otx_2x('OTX**', 10, 75, 50, 15, 100, 100),
+  otx_x('OTX*', 7, 50, 50, 10, 70, 50),
+  otx('OTX', 5, 0, 0, 7, 0, 0),
+  rtx('RTX', 4, 0, 0, 5, 0, 0),
+  okrTx('Okr. TX', 2, 0, 0, 3, 0, 0),
+  clubTx('Klub. TX', 1, 0, 0, 2, 0, 0);
 
   final String fullName;
   // ignore: non_constant_identifier_names
-  final int RT;
+  final int RT, RTu40;
   // ignore: non_constant_identifier_names
-  final int PZ;
+  final int PZ, PZu40;
+  final int aPKL, aPKLu40;
 
-  const TournamentRank(this.fullName, this.RT, this.PZ);
+  const TournamentRank(this.fullName, this.RTu40, this.PZu40, this.aPKLu40,
+      this.RT, this.PZ, this.aPKL);
+
+  int getRT(int boards) {
+    return (boards >= 40) ? RT : RTu40;
+  }
+
+  int getPZ(int boards) {
+    return (boards >= 40) ? PZ : PZu40;
+  }
+
+  int getaPKL(int boards) {
+    return (boards >= 40) ? aPKL : aPKLu40;
+  }
 
   @override
   String toString() {
@@ -26,36 +40,53 @@ enum TournamentRank {
   }
 }
 
-List<int> calculatePKL(int nofPairticipants, TournamentRank rank, double sumaWK,
-    {bool disjoinedClassification = false}) {
+List<int> calculatePKL(int numOfParticipants, int nofPlayers,
+    TournamentRank rank, double sumaWK, TournamentType type, int nofBoards) {
   List<int> res = [];
-  double PZ = max(
-      (sumaWK / nofPairticipants) * rank.RT + 0.05 * nofPairticipants,
-      rank.PZ.toDouble());
+  // ignore: non_constant_identifier_names
+  int PZ = (max(
+              max((sumaWK / nofPlayers), 0.15) * rank.getRT(nofBoards) +
+                  0.05 * nofPlayers,
+              rank.getPZ(nofBoards).toDouble()) *
+          ((type == TournamentType.teams) ? 1.25 : 1))
+      .ceil();
 
-  int twoPercentIndex = (0.02 * nofPairticipants).ceil();
-  int eighteenPercentIndex = (0.2 * nofPairticipants).ceil();
-  int fiftyPercentIndex = (0.5 * nofPairticipants).ceil();
+  double twoPercentIndex = 0.02 * numOfParticipants;
+  double twentyPercentIndex = 0.2 * numOfParticipants;
+  double fiftyPercentIndex = 0.5 * numOfParticipants;
 
-  res.add(PZ.ceil());
-  for (int place = 1; place < nofPairticipants; place++) {
-    if (place <= twoPercentIndex) {
-      res.add((0.9 * PZ).ceil());
-    } else if (place <= eighteenPercentIndex) {
-      double factor = (0.9 -
-          0.7 *
-              ((place - twoPercentIndex) /
-                  (eighteenPercentIndex - twoPercentIndex)));
-      res.add((factor * PZ).ceil());
-    } else if (place <= fiftyPercentIndex) {
-      double factor = (0.2 -
-          0.2 *
-              ((place - eighteenPercentIndex) /
-                  (fiftyPercentIndex - eighteenPercentIndex)));
-      res.add(max((factor * PZ).ceil(), 1));
+  res.add(PZ);
+
+  for (int place = 2; place <= numOfParticipants; place++) {
+    double factor = (place - 1);
+    if (factor < twoPercentIndex) {
+      res.add(((PZ * (twoPercentIndex - factor) + 0.9 * PZ * factor) /
+              twoPercentIndex)
+          .ceil());
+    } else if (factor < twentyPercentIndex) {
+      res.add(((0.9 * PZ * (twentyPercentIndex - factor) +
+                  0.2 * PZ * (factor - twoPercentIndex)) /
+              (twentyPercentIndex - twoPercentIndex))
+          .ceil());
+    } else if (factor < fiftyPercentIndex) {
+      res.add(((0.2 * PZ * (fiftyPercentIndex - factor)) /
+              (fiftyPercentIndex - twentyPercentIndex))
+          .ceil());
     } else {
       res.add(1);
     }
+  }
+
+  return res;
+}
+
+List<int> calculateCP(int numOfParticipants, TournamentType type) {
+  List<int> res = [];
+  double tournamentTypeFactor = (type == TournamentType.teams) ? 4 : 2;
+  for (int place = 1; place <= numOfParticipants; place++) {
+    res.add(((numOfParticipants - place + 1) * tournamentTypeFactor +
+            max(0, 5 - (place - 1) * 2))
+        .ceil());
   }
   return res;
 }